+++ app/src/main/java/com/moyu/browser_moyu/MainActivity.java
@@ -1,0 +1,14 @@
+package com.moyu.browser_moyu;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
