+++ app/src/main/res/layout/activity_main.xml
@@ -1,18 +1,150 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="#fafafa"
+    android:orientation="vertical"
     tools:context=".MainActivity">
 
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Hello World!"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+    <!-- 顶部地址栏 -->
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="40dp"
+        android:orientation="horizontal"
+        android:background="@drawable/top_bar_bg">
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+        <ImageView
+            android:id="@+id/webIcon"
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:contentDescription="favicon"
+            android:focusable="true"
+            android:focusableInTouchMode="true"
+            android:padding="11dp"
+            android:src="@drawable/internet"
+            android:background="#01000000"/>
+
+        <EditText
+            android:id="@+id/textUrl"
+            android:layout_width="0dp"
+            android:layout_height="34dp"
+            android:layout_weight="1"
+            android:layout_marginTop="3dp"
+            android:background="#01000000"
+            android:hint="@string/input_url"
+            android:imeOptions="actionGo"
+            android:inputType="textUri"
+            android:maxLines="1"
+            android:textSize="15sp"
+            android:textColor="#606060"
+            android:textColorHint="#9f9f9f"/>
+
+        <ImageView
+            android:id="@+id/btnStart"
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:contentDescription="refresh"
+            android:padding="11dp"
+            android:layout_marginRight="2dp"
+            android:layout_marginEnd="2dp"
+            android:src="@drawable/refresh"
+            android:background="#01000000"/>
+
+    </LinearLayout>
+
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1"
+        android:layout_marginTop="1dp"
+        android:layout_marginBottom="1dp"
+        android:background="#d1d1d1">
+
+        <WebView
+            android:id="@+id/webView"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"/>
+
+        <ProgressBar
+            android:id="@+id/progressBar"
+            android:layout_width="match_parent"
+            android:layout_height="2dp"
+            android:max="100"
+            android:progress="50"/>
+    </FrameLayout>
+
+    <!-- 底部导航栏 -->
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="40dp"
+        android:baselineAligned="false"
+        android:orientation="horizontal">
+
+        <!-- 后退按钮 -->
+        <RelativeLayout style="@style/NavItem">
+
+            <ImageView
+                android:id="@+id/goBack"
+                style="@style/NavIcon"
+                android:contentDescription="@string/nav_left"
+                android:src="@drawable/nav_left" />
+
+        </RelativeLayout>
+
+        <!-- 前进按钮 -->
+        <RelativeLayout style="@style/NavItem">
+
+            <ImageView
+                android:id="@+id/goForward"
+                style="@style/NavIcon"
+                android:contentDescription="@string/nav_right"
+                android:src="@drawable/nav_right" />
+
+        </RelativeLayout>
+
+        <!-- 菜单 -->
+        <RelativeLayout style="@style/NavItem">
+
+            <ImageView
+                android:id="@+id/navSet"
+                style="@style/NavIcon"
+                android:contentDescription="@string/nav_set"
+                android:src="@drawable/nav_set" />
+
+        </RelativeLayout>
+
+        <!-- 窗口按钮 -->
+        <RelativeLayout style="@style/NavItem">
+
+            <ImageView
+                style="@style/NavIcon"
+                android:contentDescription="@string/nav_page"
+                android:src="@drawable/nav_page" />
+
+            <TextView
+                style="@style/NavIcon"
+                android:gravity="center"
+                android:text="1"
+                android:textAlignment="center"
+                android:textColor="#555"
+                android:textSize="12sp" />
+
+        </RelativeLayout>
+
+        <!-- 主页按钮 -->
+        <RelativeLayout style="@style/NavItem">
+
+            <ImageView
+                android:id="@+id/goHome"
+                style="@style/NavIcon"
+                android:contentDescription="@string/nav_home"
+                android:src="@drawable/nav_home" />
+
+        </RelativeLayout>
+
+
+    </LinearLayout>
+</LinearLayout>
