+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,10 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <!-- 下拉到底的阴影 -->
+    <color name="colorPrimary">#bababa</color>
+
+    <!-- 导航栏颜色 -->
+    <color name="colorPrimaryDark">#3d3d3d</color>
+    <color name="colorAccent">#9b9b9b</color>
 </resources>
