+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,15 @@
 <resources>
-    <string name="app_name">browse_moyu</string>
+    <string name="app_name">browser_moyu</string>
+
+    <string name="input_url">搜索或输入网址</string>
+
+    <!-- 网站首页 -->
+    <string name="home_url">https://cn.bing.com/</string>
+
+    <!-- 底部导航 -->
+    <string name="nav_right">前进</string>
+    <string name="nav_left">后退</string>
+    <string name="nav_set">设置</string>
+    <string name="nav_page">窗口</string>
+    <string name="nav_home">主页</string>
 </resources>
