+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,4 @@
-#Mon Jul 12 15:54:08 CST 2021
+#Wed Jul 14 18:30:46 CST 2021
 distributionBase=GRADLE_USER_HOME
 distributionUrl=https\://services.gradle.org/distributions/gradle-6.7.1-bin.zip
 distributionPath=wrapper/dists
