+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = "browse_moyu"
+rootProject.name = "browser_moyu"
 include ':app'
